SearchDatabase/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
aniso8601==9.0.1
bs4==0.0.1
Flask==2.0.2
Flask-RESTful==0.3.9
gnews
ibm-cloud-sdk-core==3.13.2
ibm-watson==5.3.0
mysql-connector-python==8.0.28
newspaper3k==0.2.8
numpy==1.21.4
pandas==1.4.1
tqdm==4.62.3
tweepy==4.5.0
urllib3==1.26.7
pyodbc